pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-medical_record_abstraction_env"
version = "0.1.0"
description = "Medical Record Abstraction environment for OpenEnv — clinical NLP with 3 difficulty tasks"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "openai>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "medical_record_abstraction_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["medical_record_abstraction_env", "medical_record_abstraction_env.server", "medical_record_abstraction_env.data"]
package-dir = { "medical_record_abstraction_env" = ".", "medical_record_abstraction_env.server" = "server", "medical_record_abstraction_env.data" = "data" }